requirements-dev.txt
flake8==3.7.8
moto==1.3.13
pytest==5.2.0
pytest-cov==2.7.1
freezegun==0.3.12
requests-mock[fixture]==1.7.0
black==19.10b0
prometheus_client==0.7.1